+++ api/src/modules/user/user.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { UserService } from './user.service';
-import { UserController } from './user.controller';
 import { PrismaModule } from 'src/prisma/prisma.module';
+import { UserController } from './presentation/user.controller';
+import { UserService } from './application/user.service';
 
 @Module({
   imports: [PrismaModule],
